requirements.txt
cffi==1.7.0
cryptography==1.5
idna==2.1
paramiko==2.0.2
pyasn1==0.1.9
pycparser==2.14
python-decouple==3.0
six==1.10.0
